+++ src/components/List/List.js
@@ -1,27 +1,47 @@
 import PropTypes from 'prop-types';
 import {createElement} from 'react';
-import styled from 'styled-components';
+import styled, {css} from 'styled-components';
 
-import * as structureCss from 'src/styles/structure';
+import * as listCss from 'src/components/List/internal/listCss';
 
 export const styleTypeStyles = (props) => {
-  if (!props.styleType) {
+  if (props.styleType !== 'plain') {
     return '';
   }
 
-  return `
+  return css`
     list-style: none;
     padding: 0;
   `;
 };
 
-export const baseContainerStyles = (props) => {
-  return `
-    & > & {
-      padding: 0 0 0 ${structureCss.spacing.small};
+export const nestedStyles = (props) => {
+  let marginLeft = 0;
+
+  if (props.isExpandable) {
+    marginLeft = `calc(calc(${listCss.variables.paddingLeft} + ${listCss.variables.expandableIcon.size}) * -1)`;
+  }
+
+  return css`
+    & & {
+      margin-left: ${marginLeft};
+      padding: 0 0 0 ${listCss.variables.paddingLeft};
     }
 
-    ${styleTypeStyles(props)}
+    & & li:first-child {
+      padding-top: 0;
+    }
+
+    & & li:last-child {
+      padding-bottom: 0;
+    }
+  `;
+};
+
+export const baseContainerStyles = (props) => {
+  return css`
+    line-height: 1.6rem;
+    ${nestedStyles(props)} ${styleTypeStyles(props)};
   `;
 };
 
@@ -30,11 +50,13 @@ export const List = styled(({tag, children, styleType, ...props}) => createEleme
 `;
 
 List.propsTypes = {
+  isExpandable: PropTypes.bool,
   styleType: PropTypes.oneOf(['plain']),
   tag: PropTypes.oneOf(['ol', 'ul']),
 };
 
 List.defaultProps = {
+  isExpandable: false,
   styleType: null,
   tag: 'ul',
 };