+++ web/app/react/components/MainNavigationSection.js
@@ -1,9 +1,43 @@
 import PropTypes from 'prop-types';
 import React from 'react';
+import {connect} from 'react-redux';
+import styled, {css} from 'styled-components';
 
-import ExpandableList from 'src/components/List/ExpandableList';
-import ListItem from 'src/components/List/ListItem';
-import SvgIcon from 'src/components/SvgIcon/SvgIcon';
+import * as structureCss from 'src/styles/structure';
+import * as themesCss from 'src/styles/themes';
+
+import {List, ExpandableListRenderer, ExpandableListHeader, ListItem, SvgIcon} from 'src';
+
+export const listItemStyles = (props) => {
+  return css`
+    cursor: pointer;
+    ${props.isActive ? 'font-weight: bold;' : ''} &:hover {
+      background: ${themesCss.light.application.activeState};
+    }
+  `;
+};
+
+const ListItemStyled = styled(ListItem)`
+  ${listItemStyles};
+`;
+
+export const expandableListStyles = () => {
+  const padding = structureCss.spacing.extraSmall;
+
+  return css`
+    padding: ${padding} ${padding} 0 ${padding};
+  `;
+};
+
+const ListStyled = styled(List)`
+  ${expandableListStyles};
+`;
+
+export const mapStateToProps = ({router}) => {
+  return {
+    router,
+  };
+};
 
 class MainNavigationSection extends React.Component {
   getCssClasses() {
@@ -22,7 +56,21 @@ class MainNavigationSection extends React.Component {
     }
   };
 
-  renderItems() {
+  navigationRenderer = ({isExpanded, onToggleList}) => {
+    return (
+      <ListStyled styleType="plain">
+        <ExpandableListHeader
+          isExpanded={isExpanded}
+          onClick={onToggleList}
+        >
+          {this.props.headerNode}
+        </ExpandableListHeader>
+        {this.renderItems(isExpanded)}
+      </ListStyled>
+    );
+  };
+
+  renderItems(isExpanded) {
     let itemNodes = [];
 
     if (this.props.items.length > 0) {
@@ -34,15 +82,17 @@ class MainNavigationSection extends React.Component {
         }
 
         itemNodes.push(
-          <ListItem
+          <ListItemStyled
             data-to={item.to}
+            isActive={this.props.router.location.pathname === item.to}
             isExpandable
+            isExpanded={isExpanded}
             key={item.id}
             onClick={this.onClickMenuItem}
           >
             {svgIconNode}
             {item.display}
-          </ListItem>
+          </ListItemStyled>
         );
       });
     }
@@ -51,7 +101,13 @@ class MainNavigationSection extends React.Component {
   }
 
   render() {
-    return <ExpandableList handleNode={this.props.headerNode}>{this.renderItems()}</ExpandableList>;
+    return (
+      // eslint-work-around
+      <ExpandableListRenderer
+        initiallyExpanded
+        render={this.navigationRenderer}
+      />
+    );
   }
 }
 
@@ -59,6 +115,7 @@ MainNavigationSection.propTypes = {
   className: PropTypes.string,
   headerNode: PropTypes.node.isRequired,
   items: PropTypes.array.isRequired,
+  router: PropTypes.object.isRequired,
 };
 
 MainNavigationSection.contextTypes = {
@@ -71,4 +128,4 @@ MainNavigationSection.defaultProps = {
   items: [],
 };
 
-export default MainNavigationSection;
+export default connect(mapStateToProps)(MainNavigationSection);